backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
Faker==37.8.0
sqlparse==0.5.3
tzdata==2025.2
